+++ HISTORY.rst
@@ -3,13 +3,23 @@ History
 =======
 
 -------------------
+1.0.6 (2023-06-14)
+-------------------
+
+* The CLI command run-verification-tests now required three inputs (capif registration file, certificates folder and verification file) to launch NEF and TSN verification pipelines.
+* CAPIFProviderConnector has a new method offboard_and_deregister_netapp() that allows a network app to be offboarded and deregistered from CAPIF.
+* The SDK classes LocationSubscriber,ConnectionMonitor,QosAwareness don't accept the parameter 'nef_bearer_access_token' any more.
+* The CLI method register_and_onboard_to_capif now accepts a second parameter --environment that takes values "production" or "development".
+If this parameter is set to production then validation tests are running in the background, just after the registration and onboarding of the network app finishes.
+
+-------------------
 1.0.5 (2023-05-24)
 -------------------
 * Bug fix on CAPIFProviderConnector. When storing the capif_cert_server.pem the port was hardcoded.
 * Improvement on verification pipelines. Required user and password.
 * TSN verification pipeline added.
 * Validation pipeline included.
- 
+
 -------------------
 1.0.4 (2023-04-27)
 -------------------